requirements.txt
discord.py >= 1.7.0
pymongo >= 3.11.3
pymongo[srv] >= 3.11.3
APScheduler >= 3.7.0
PyNaCl >= 1.4.0